# src/IntakeGate.Web/openapi/intake-gate.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "IntakeGate.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080/"
    }
  ],
  "paths": {
...
    "/api/home/summary": {
      "get": {
        "tags": [
          "IntakeGate.Host"
        ],
        "parameters": [
          {
            "name": "windowDays",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "required": [
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "EngineeringReadyRateResponse": {
        "required": [
          "numerator",
          "denominator",
          "percentage"
        ],
        "type": "object",
        "properties": {
          "numerator": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "denominator": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "percentage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "EstimatedAiCostSummaryResponse": {
        "required": [
          "amount",
          "currency",
          "evaluationsWithEstimate",
          "evaluationsWithoutEstimate",
          "complete"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "evaluationsWithEstimate": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "evaluationsWithoutEstimate": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "complete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EstimatedCostResponse": {
        "required": [
          "amount",
          "currency",
          "pricingIdentity"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "pricingIdentity": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ExecutionMode": {
        "enum": [
          "live",
          "dryRun"
        ]
      },
      "HomeHealthWarningResponse": {
        "required": [
          "code",
          "title",
          "message",
          "detailUrl"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailUrl": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "HomeSummaryResponse": {
        "required": [
          "windowDays",
          "generatedAtUtc",
          "windowStartInclusiveUtc",
          "windowEndExclusiveUtc",
          "evaluatedCount",
          "engineeringReadyCount",
          "intakeIncompleteCount",
          "errorCount",
          "notEligibleCount",
          "skippedCount",
          "duplicateUpdatesSuppressedCount",
          "engineeringReadyRate",
          "estimatedAiCost",
          "recentRuns",
          "healthWarnings"
        ],
        "type": "object",
        "properties": {
          "windowDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "generatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "windowStartInclusiveUtc": {
            "type": "string",
            "format": "date-time"
          },
          "windowEndExclusiveUtc": {
            "type": "string",
            "format": "date-time"
          },
          "evaluatedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "engineeringReadyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "intakeIncompleteCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "errorCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "notEligibleCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "skippedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "duplicateUpdatesSuppressedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "engineeringReadyRate": {
            "$ref": "#/components/schemas/EngineeringReadyRateResponse"
          },
          "estimatedAiCost": {
            "$ref": "#/components/schemas/EstimatedAiCostSummaryResponse"
          },
          "recentRuns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunSummaryResponse"
            }
          },
          "healthWarnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HomeHealthWarningResponse"
            }
          }
        },
        "additionalProperties": false
      },
...
      "OperationalActorResponse": {
        "required": [
          "type",
          "userId",
          "username"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/OperationalActorType"
          },
          "userId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "username": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "OperationalActorType": {
        "enum": [
          "user",
          "system",
          "historicalUnknown"
        ]
      },
...
      "OperationalRunStatus": {
        "enum": [
          "running",
          "completed",
          "completedWithErrors",
          "error"
        ]
      },
...
      "RunSummaryResponse": {
        "required": [
          "runId",
          "invocationType",
          "triggeredBy",
          "startedAtUtc",
          "completedAtUtc",
          "durationMilliseconds",
          "effectiveMode",
          "status",
          "profileId",
          "configurationGenerationId",
          "ticketsDiscovered",
          "ticketsEvaluated",
          "engineeringReadyCount",
          "intakeIncompleteCount",
          "errorCount",
          "notEligibleCount",
          "skippedCount",
          "duplicateUpdatesSuppressedCount",
          "azureDevOpsMutationCount",
          "tokenUsage",
          "estimatedCost",
          "errors"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "invocationType": {
            "$ref": "#/components/schemas/RunTriggerType"
          },
          "triggeredBy": {
            "$ref": "#/components/schemas/OperationalActorResponse"
          },
          "startedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "durationMilliseconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "effectiveMode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "status": {
            "$ref": "#/components/schemas/OperationalRunStatus"
          },
          "profileId": {
            "type": "string"
          },
          "configurationGenerationId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "ticketsDiscovered": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ticketsEvaluated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "engineeringReadyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "intakeIncompleteCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "errorCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "notEligibleCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "skippedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "duplicateUpdatesSuppressedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "azureDevOpsMutationCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tokenUsage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TokenUsageResponse"
              }
            ]
          },
          "estimatedCost": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EstimatedCostResponse"
              }
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SafeOperationalErrorResponse"
            }
          }
        },
        "additionalProperties": false
      },
...
      "RunTriggerType": {
        "enum": [
          "scheduled",
          "manualIncremental",
          "manualWorkItem"
        ]
      },
      "SafeOperationalErrorResponse": {
        "required": [
          "category",
          "message"
        ],
        "type": "object",
        "properties": {
          "category": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "TokenUsageResponse": {
        "required": [
          "inputTokens",
          "outputTokens",
          "totalTokens"
        ],
        "type": "object",
        "properties": {
          "inputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "outputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}